+++ assets/javascripts/kitten/components/cards/contribution-card/components/input.js
@@ -3,25 +3,34 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import { TextInputWithUnit } from '../../../form/text-input-with-unit'
 
-export const Input = props => {
+export const Input = ({ valid, onChange, className, ...props }) => {
   const [isInputEmpty, setEmptyInput] = useState(true)
 
   return (
     <TextInputWithUnit
       wrapperProps={{
-        className: classNames(
-          { 'k-ContributionCard__inputWrapper--isEmpty': isInputEmpty },
-        ),
+        className: classNames({
+          'k-ContributionCard__inputWrapper--isEmpty': isInputEmpty,
+        }),
       }}
       variant="orion"
-      {...props}
-      className={classNames(
-        'k-ContributionCard__input',
-        props.className,
-      )}
-      onChange={(event) => {
+      valid={valid}
+      className={classNames('k-ContributionCard__input', className)}
+      onChange={event => {
         setEmptyInput(event.target?.value?.length === 0)
+        onChange(event)
       }}
+      {...props}
     />
   )
+}
+
+Input.propTypes = {
+  valid: PropTypes.bool,
+  onChange: PropTypes.func,
+  className: PropTypes.string,
+}
+
+Input.defaultProps = {
+  onChange: () => null,
 }